+++ encryption/blowfish.c
@@ -39,7 +39,7 @@ ccrypto_error_type ccrypto_blowfish_encrypt(const uint8_t *key,
     BF_KEY bf_key;
     BF_set_key(&bf_key, (int)key_size, key);
 
-    for (int i = 0; i < data_size; i += BLOCK_SIZE)
+    for (size_t i = 0; i < data_size; i += BLOCK_SIZE)
     {
         // Encrypt the data
         BF_ecb_encrypt(data, output, &bf_key, BF_ENCRYPT);
